GeneratePackages: Connect the parsing of array declaration on fields to to the support in the writer and introduce arrayEaters.
The arrayEaters are basic field types that "eats" a level of array declarations. This is done by having the generated packet skip the last level of arrays and let the field handle it. Create a packet with an array of strings and a string to test that fields of arrayEaters are handled correctly. Update tests data and hard coded data.

The connection is done by having PacketsStore pass on array declaration information to the fields.