GeneratePackets: Bugfix: Read constants from C that previously were unparseable.
The old code for deciding if a line was a define had bugs. Remove it and set a Boolean flag while parsing EOL-sensitive code in stead. This is safe as long as no item caring about EOL are inside another item caring about EOL. Since no #define is inside a #define this should be OK.