GeneratePackets: Java Generator: Typed code: Add hints to IR.
A hint permits an area of code to be marked as special by the HasAtoms that generate it. A hint is began before a code atom and ended after a code atom. It should be ended to not mess up following code. Only the most recent unended hint is exposed to code style rules via ScopeInfo at the moment.