GeneratePackets: Java Generator: Formating: CodeStyleBuilder: Focus hints on the space between two items as that is where the pointer is.
Change IR semantic to be more consistant. The hints are now focused on the exact middle between two code atoms. This mean than hints that were finished in the left element and hint began in the rigth elements aren't there when the rules run. (In the old system the hints dtarted in the next element were present so the pointer cuold be said to be a bit to the right)