Bug fix: GeneratePackets: Make generated basic field type STRING measure its size in bytes.
This fixes the current problem of non ascii letters in a STRING in a packet causing the failure of reading it from the network. Before STRING was generated using TerminatedArray it manifested it self as non ASCII letters being displayed as two different letters. Only support using the systems default Charset for now.