BindingsUsers: Add new program ProxyRecorder that should lay between the Freeciv client and server and record the packets sendt and recieved.
The motivation for this is to record traces of Freeciv games and to test the protocol implementation. Recorded traces can be used to train AI's and let humans view games later from the point of view of the recorded player(s). At the moment all it does is acceptiong connections from clients and print the packets they send to the standard output.