GeneratePackets: enteties: Struct: Make it possible to generate a basic field type from a struct under some conditions.
This can't generate the two structs used in packets.def as packets.def refer to functions in dataio.c in stead of telling how much data each field of a Struct should read.

The basic field type is generated by a maker. The requirement asking it to start specifies a NetworkIO for each field using the syntax introduced in the previous commit. It then requires an anonymous field type alias for each field.