Bug fix: FreecivRecorder: PlayToServer: Don't assume to hace received the challenge file name when the request for hack access was sent the last time.
To do this there has to be two plumbing threads. One read from the trace and send to the server. The other recieves from the server and look for the challenge file name. If the trace to server thread discovers that no challenge file name is there when it wants to send it it will wait for it.