GeneratePackets: Make it easier to comply with the GPL when distributing by adding a copy of the Freeciv sources generated from to the generated Java code.
The GPLv2 section 3 requires that the source code it self or a written offer to give it to anyone accompanies the distribution of "object code or executable form". The source code is defined as "the preferred form of the work for making modifications to it." The Freeciv C source can therefore be said to be the source code of the generated Java code unless the Java code has been manually modified after it was generated. Remembering to include it or an offer for it can be hard.

To solve this GeneratePackets copies the Freeciv sources used to generate the Java code and put them next to the generated Java code if the option --gpl-source is set to true.