Bug fix: Generated packets: Packets: Generate the from values factory for field less packets.
When creation from values was a constructor it had no name. It would therefore end up as a duplicate of the constructor from fields. As the constructor from fields existed it served as a constructor from values as well. When it got a name (as a factory) it should have been generated.