Bug fix: Core: utility: EndsInEternalZero: Fix the range of read() so it is 0 to 255
Java's byte is signed. It's range is therefore -128 to 127. EndsInEternalZero takes the first bytes as an array of byte. The range of values InputStream's read() is allowed to return is 0 to 255. (-1 signals that the end is reached) Change the values from the byte array to the correct range before returning them in read().