GeneratePackets: enteties: Field: Make packet fields protected by default.
This way the common fields can be shared among variants in an abstract parent class so the parent can be used when the capability modified fields aren't needed. Changes from new capabilities will cause compile time errors in code that use the parent and don't support the new variant in stead of silently doing nothing.