Bug fix: GeneratePackets: ignore multiline C comments as white space when asked to treat comments as white space.
A C constant followed by a multiline comment will be unparseable if this isn't done. Freeciv trunk has a constant like this in a constant needed on the Java side from svn commit 23881.

This change makes the parser ignore line breaks in C comments even if it is told to leave line breaks alone. Doing this can be rationalized since it was asked to threat C comments as white space and a C comment may be multiline.

The implementation of regular expressions in the reference Java standard library use recursion. The recursion depend on the input size, not the regex size. The recursion isn't optimized away by the compiler or by the JVM. Running out of stack when processing large comments will probably become an issue.