DependencyHandler: Add BlameShifter - a Maker that has blame information.
A Maker may be able to make an educated guess about why requirements, like its input, is missing. An example is a maker that know what items its input is made from.

Add a new interface, BlameShifter, to let the DependencyStore know that the information is there. Make DependencyStore add the blame information when it detects that an added Maker is a BlameShifter.